MainMenu input

+++ GameEngine/Animation.h
@@ -8,7 +8,7 @@ public:
 	~Animation();
 	void UpdateAnimation(double x, double y, SDL_RendererFlip flip = SDL_FLIP_NONE);
 	int CURRENT_FRAME = 0;
-	int scale = 1;
+	double scale = 1;
 	int total_frames;
 	int WALKING_ANIMATION_FRAMES;
 	std::vector<SDL_Rect> gSpriteClips;
